requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
scikit-learn
matplotlib
seaborn
plotly
streamlit
httpx
openai
pytest
black
isort
python-multipart
langchain-community
langchain-core
langchain-experimental
langchainhub
langchain-google-genai
langgraph
pinecone
langchain_huggingface
sentence-transformers
tqdm
pymongo[srv]==3.11
streamlit-extras
